hold a storage process as a child always
it is linked to us, so if it crashes the worker crashes, and the
worker pool catches that. makes things a lot simpler as the worker
now always has access to a storage item. as the workers are long-lived,
so are the storage (and therefore riak connections) preserving
the vector clocks as previously